requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Mail==0.9.1
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
